Start splitting out data access and domain layers
