+++ AcmeStudioRefactorTest_IF.sln
@@ -3,7 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.6.33815.320
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AcmeStudios.ApiRefactor", "AcmeStudios.ApiRefactor\AcmeStudios.ApiRefactor.csproj", "{DA45EFF5-471F-4270-B94F-C10DC04FFBC8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AcmeStudios.ApiRefactor", "AcmeStudios.ApiRefactor\AcmeStudios.ApiRefactor.csproj", "{DA45EFF5-471F-4270-B94F-C10DC04FFBC8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AcmeStudios.ApiRefactor.DataAccess", "AcmeStudios.ApiRefactor.DataAccess\AcmeStudios.ApiRefactor.DataAccess.csproj", "{00EE5F38-43D4-4574-84E2-8D09CFD79143}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AcmeStudios.ApiRefactor.Domain", "AcmeStudios.ApiRefactor.Domain\AcmeStudios.ApiRefactor.Domain.csproj", "{9F5D7904-AD86-480D-A930-785EB870A604}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +19,14 @@ Global
 		{DA45EFF5-471F-4270-B94F-C10DC04FFBC8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DA45EFF5-471F-4270-B94F-C10DC04FFBC8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DA45EFF5-471F-4270-B94F-C10DC04FFBC8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{00EE5F38-43D4-4574-84E2-8D09CFD79143}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{00EE5F38-43D4-4574-84E2-8D09CFD79143}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{00EE5F38-43D4-4574-84E2-8D09CFD79143}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{00EE5F38-43D4-4574-84E2-8D09CFD79143}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9F5D7904-AD86-480D-A930-785EB870A604}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9F5D7904-AD86-480D-A930-785EB870A604}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9F5D7904-AD86-480D-A930-785EB870A604}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9F5D7904-AD86-480D-A930-785EB870A604}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
