+++ AcmeStudios.ApiRefactor.DataAccess/appsettings.json
@@ -1,0 +1,14 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "StudioConnection": "Server=localhost;Database=Studio;Trusted_Connection=True;MultipleActiveResultSets=true"
+
+  }
+}
