+++ AcmeStudios.ApiRefactor.Domain/StudioItemType.cs
@@ -1,0 +1,16 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Text.Json.Serialization;
+
+namespace AcmeStudios.ApiRefactor.Domain
+{
+    public class StudioItemType
+    {
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        public int StudioItemTypeId { get; set; }
+        [Required]
+        public string Value { get; set; }
+        [JsonIgnore]
+        public ICollection<StudioItem> StudioItem { get; set; }
+    }
+}
