+++ AcmeStudios.ApiRefactor/AcmeStudios.ApiRefactor.csproj
@@ -24,5 +24,10 @@
     <Folder Include="wwwroot\" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\AcmeStudios.ApiRefactor.DataAccess\AcmeStudios.ApiRefactor.DataAccess.csproj" />
+    <ProjectReference Include="..\AcmeStudios.ApiRefactor.Domain\AcmeStudios.ApiRefactor.Domain.csproj" />
+  </ItemGroup>
+
 
 </Project>
