+++ AcmeStudios.ApiRefactor/DTOs/StudioItemTypeDto.cs
@@ -1,0 +1,8 @@
+﻿namespace AcmeStudios.ApiRefactor.DTOs
+{
+    public class StudioItemTypeDto
+    {
+        public int StudioItemTypeId { get; set; }
+        public string Value { get; set; }
+    }
+}
