+++ AcmeStudios.ApiRefactor/appsettings.json
@@ -1,12 +1,4 @@
 {
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "AllowedHosts": "*",
   "ConnectionStrings": {
     "StudioConnection": "Server=localhost;Database=Studio;Trusted_Connection=True;MultipleActiveResultSets=true"
